refactor: move components to their right place
