+++ web-react-ts/src/pages/campaigns/components/cards/DefaulterDetailsCard.tsx
