+++ web-react-ts/src/pages/campaigns/equipment/council/CouncilByConstituencyEquipmentDefaulters.tsx
@@ -6,7 +6,7 @@ import { ChurchContext } from 'contexts/ChurchContext'
 import { COUNCIL_BY_CONSTITUENCY_EQUIPMENT_DEFAULTERS } from 'pages/campaigns/CampaignQueries'
 import DefaulterDetailsCard, {
   EquipmentDefaulters,
-} from 'pages/campaigns/components/buttons/DefaulterDetailsCard'
+} from 'pages/campaigns/components/cards/DefaulterDetailsCard'
 import React, { useContext } from 'react'
 import { Container, Row } from 'react-bootstrap'
