+++ web-react-ts/src/pages/campaigns/equipment/gathering-service/GatheringServiceByStreamEquipmentDefaulters.tsx
@@ -7,7 +7,7 @@ import { MemberContext } from 'contexts/MemberContext'
 import { GATHERING_SERVICE_BY_STREAM_EQUIPMENT_DEFAULTERS } from 'pages/campaigns/CampaignQueries'
 import DefaulterDetailsCard, {
   EquipmentDefaulters,
-} from 'pages/campaigns/components/buttons/DefaulterDetailsCard'
+} from 'pages/campaigns/components/cards/DefaulterDetailsCard'
 import React, { useContext } from 'react'
 import { Container, Row } from 'react-bootstrap'
